Initial drawing code; change emscripten output to js and use our own html.
JS drawing object is now a global. Emscripten doesn't support passing JS objects through C void* handles. They appear to work for a while (while they're cached in register vars), but get lost when stored in the heap (which is an Int32Array). Global drawing handle is a simple workaround for now.